institutions_design_poc/requirements.txt
pymysql>=1.1.0
streamlit>=1.40.0
httpx>=0.28.0
pydantic>=2.0
pandas>=2.0
openpyxl>=3.1.0
python-dotenv>=1.0
openai>=1.0
tqdm>=4.0
numpy>=2.0
